webauthn.go
package webauthn

import (
	"crypto/sha256"
	"encoding/base64"
	"errors"
	"fmt"
)

// PublicKeyCredential represents the credential created during registration or used during authentication.
type PublicKeyCredential struct {
    ID    string
    RawID []byte
    Type  string
    Response AuthenticatorResponse
}

// AuthenticatorResponse contains either the authenticator attestation or assertion response.
type AuthenticatorResponse struct {
    AttestationResponse *AuthenticatorAttestationResponse
    AssertionResponse   *AuthenticatorAssertionResponse
}

// AuthenticatorAttestationResponse represents the attestation response during registration.
type AuthenticatorAttestationResponse struct {
    ClientDataJSON    string
    AttestationObject string
}

// AuthenticatorAssertionResponse represents the assertion response during authentication.
type AuthenticatorAssertionResponse struct {
    ClientDataJSON    string
    AuthenticatorData string
    Signature         string
    UserHandle        string
}

// ParseAttestationResponse parses the attestation response from a registration ceremony.
func ParseAttestationResponse(attResp AuthenticatorAttestationResponse) (parsedData map[string]interface{}, err error) {
    // Decode the attestation object
    attestationObject, err := base64.StdEncoding.DecodeString(attResp.AttestationObject)
    if err != nil {
        return nil, fmt.Errorf("failed to decode attestation object: %w", err)
    }

    // Further processing would be required to fully parse the attestation object
    // (CBOR decoding, etc.). This is a simplified example.
    parsedData = map[string]interface{}{
        "attestationObject": attestationObject,
    }

    return parsedData, nil
}

// VerifyAttestation verifies the attestation data during registration.
func VerifyAttestation(parsedData map[string]interface{}, clientDataJSON string) error {
    // This function would need to perform a series of steps to verify the attestation,
    // such as verifying signatures, checking certificates, etc.
    // This is a simplified example.

    // Extract the attestation object
    attestationObject, ok := parsedData["attestationObject"].([]byte)
    if !ok {
        return errors.New("invalid attestation object")
    }

    // Compute the hash of the client data
    clientDataHash := sha256.Sum256([]byte(clientDataJSON))

    // Further verification steps...

    fmt.Printf("Client Data Hash: %x\n", clientDataHash)
    fmt.Printf("Attestation Object: %x\n", attestationObject)

    return nil
}

// ParseAssertionResponse parses the assertion response from an authentication ceremony.
func ParseAssertionResponse(assertResp AuthenticatorAssertionResponse) (parsedData map[string]interface{}, err error) {
    // Decode the authenticator data
    authData, err := base64.StdEncoding.DecodeString(assertResp.AuthenticatorData)
    if err != nil {
        return nil, fmt.Errorf("failed to decode authenticator data: %w", err)
    }

    // Further processing would be required to fully parse the authenticator data
    parsedData = map[string]interface{}{
        "authenticatorData": authData,
    }

    return parsedData, nil
}

// VerifyAssertion verifies the assertion data during authentication.
func VerifyAssertion(parsedData map[string]interface{}, clientDataJSON, signature string) error {
    // This function would need to perform a series of steps to verify the assertion,
    // such as verifying signatures, checking counters, etc.
    // This is a simplified example.

    // Extract the authenticator data
    authData, ok := parsedData["authenticatorData"].([]byte)
    if !ok {
        return errors.New("invalid authenticator data")
    }

    // Compute the hash of the client data
    clientDataHash := sha256.Sum256([]byte(clientDataJSON))

    // Decode the signature
    sig, err := base64.StdEncoding.DecodeString(signature)
    if err != nil {
        return fmt.Errorf("failed to decode signature: %w", err)
    }

    // Further verification steps...

    fmt.Printf("Client Data Hash: %x\n", clientDataHash)
    fmt.Printf("Authenticator Data: %x\n", authData)
    fmt.Printf("Signature: %x\n", sig)

    return nil
}
